chore(apigatewayv2): refactor types out of common/base
Move types from common/base.ts and into their corresponding files -
specifically common/api.ts and common/stage.ts.

Better code organization.
