+++ packages/@aws-cdk/aws-apigatewayv2/lib/common/api.ts
@@ -1,5 +1,6 @@
 import * as cloudwatch from '@aws-cdk/aws-cloudwatch';
-import { IResource } from '@aws-cdk/core';
+import { IResource, Resource } from '@aws-cdk/core';
+import { IntegrationCache } from '../private/integration-cache';
 
 /**
  * Represents a API Gateway HTTP/WebSocket API
@@ -23,4 +24,26 @@ export interface IApi extends IResource {
    * @default - average over 5 minutes
    */
   metric(metricName: string, props?: cloudwatch.MetricOptions): cloudwatch.Metric;
+}
+
+/**
+ * Base class representing an API
+ * @internal
+ */
+export abstract class ApiBase extends Resource implements IApi {
+  abstract readonly apiId: string;
+  abstract readonly apiEndpoint: string;
+  /**
+   * @internal
+   */
+  protected _integrationCache: IntegrationCache = new IntegrationCache();
+
+  public metric(metricName: string, props?: cloudwatch.MetricOptions): cloudwatch.Metric {
+    return new cloudwatch.Metric({
+      namespace: 'AWS/ApiGateway',
+      metricName,
+      dimensions: { ApiId: this.apiId },
+      ...props,
+    }).attachTo(this);
+  }
 }
