+++ packages/@aws-cdk/aws-apigatewayv2/lib/common/stage.ts
@@ -1,5 +1,7 @@
 import { Metric, MetricOptions } from '@aws-cdk/aws-cloudwatch';
-import { IResource } from '@aws-cdk/core';
+import { IResource, Resource } from '@aws-cdk/core';
+import { IApi } from './api';
+import { ApiMapping } from './api-mapping';
 import { IDomainName } from './domain-name';
 
 /**
@@ -70,3 +72,38 @@ export interface StageAttributes {
    */
   readonly stageName: string;
 }
+
+/**
+ * Base class representing a Stage
+ * @internal
+ */
+export abstract class StageBase extends Resource implements IStage {
+  public abstract readonly stageName: string;
+  protected abstract readonly baseApi: IApi;
+
+  /**
+   * The URL to this stage.
+   */
+  abstract get url(): string;
+
+  /**
+   * @internal
+   */
+  protected _addDomainMapping(domainMapping: DomainMappingOptions) {
+    new ApiMapping(this, `${domainMapping.domainName}${domainMapping.mappingKey}`, {
+      api: this.baseApi,
+      domainName: domainMapping.domainName,
+      stage: this,
+      apiMappingKey: domainMapping.mappingKey,
+    });
+    // ensure the dependency
+    this.node.addDependency(domainMapping.domainName);
+  }
+
+  public metric(metricName: string, props?: MetricOptions): Metric {
+    return this.baseApi.metric(metricName, props).with({
+      dimensions: { ApiId: this.baseApi.apiId, Stage: this.stageName },
+    }).attachTo(this);
+  }
+}
+
