+++ packages/@aws-cdk/aws-apigatewayv2/lib/http/api.ts
@@ -2,9 +2,7 @@ import { Metric, MetricOptions } from '@aws-cdk/aws-cloudwatch';
 import { Duration } from '@aws-cdk/core';
 import { Construct } from 'constructs';
 import { CfnApi, CfnApiProps } from '../apigatewayv2.generated';
-import { IApi } from '../common/api';
-import { ApiBase } from '../common/base';
-import { DomainMappingOptions, IStage } from '../common/stage';
+import { ApiBase, DomainMappingOptions, IApi, IStage } from '../common';
 import { IHttpRouteAuthorizer } from './authorizer';
 import { IHttpRouteIntegration, HttpIntegration, HttpRouteIntegrationConfig } from './integration';
 import { BatchHttpRouteOptions, HttpMethod, HttpRoute, HttpRouteKey } from './route';
