+++ packages/@aws-cdk/aws-apigatewayv2/lib/http/stage.ts
@@ -2,9 +2,7 @@ import { Metric, MetricOptions } from '@aws-cdk/aws-cloudwatch';
 import { Stack } from '@aws-cdk/core';
 import { Construct } from 'constructs';
 import { CfnStage } from '../apigatewayv2.generated';
-import { StageOptions, IStage, StageAttributes } from '../common';
-import { IApi } from '../common/api';
-import { StageBase } from '../common/base';
+import { IApi, StageBase, StageOptions, IStage, StageAttributes } from '../common';
 import { IHttpApi } from './api';
 
 const DEFAULT_STAGE_NAME = '$default';
